Create model for certificate and improve variable naming
Previously, in order to verify the server's certificate, client must know the content of the certificate and the public key used to verify it before the server actually send the certificate.
To improve the flexibility, now whenever the client requested the server's certificate, the server will send the content, signature, and the public key instead.